Move RibbonScreenShotWindow
