--- ribbon/tests/__init__.py
@@ -1,1 +1,0 @@
-from .screenshotwindow import RibbonScreenShotWindow
